# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Indian Railways LocoNet Fault Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  createdAt: string;
  lastLogin?: string | null;
}

export interface LoginResult {
  token: string;
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserInput {
  username: string;
  email: string;
  password: string;
  role: UserInputRole;
}

export interface Fault {
  id: number;
  uniqueFaultId: string;
  alertType?: string | null;
  locoNo?: string | null;
  locoType?: string | null;
  shed?: string | null;
  zone?: string | null;
  loggedTimestamp?: string | null;
  faultCode?: string | null;
  faultDescription?: string | null;
  basicUnit?: string | null;
  category?: string | null;
  moduleName?: string | null;
  coachNumber?: string | null;
  location?: string | null;
  datapack?: string | null;
  coachType?: string | null;
  recoveryStatus?: string | null;
  downloadStatus?: string | null;
  apiTimestamp?: string | null;
  dbTimestamp?: string | null;
  createdAt: string;
}

export interface FaultListResponse {
  data: Fault[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export interface DashboardStats {
  totalFaultsToday: number;
  categoryA: number;
  categoryB: number;
  categoryC: number;
  recoveredFaults: number;
  activeFaults: number;
  todayCoaches: number;
  todayLocoCount: number;
  lastUpdated: string;
}

export interface ChartDataPoint {
  label: string;
  value: number;
}

export interface TrendDataPoint {
  timestamp: string;
  count: number;
}

export type AlertLogStatus = typeof AlertLogStatus[keyof typeof AlertLogStatus];


export const AlertLogStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export interface AlertLog {
  id: number;
  faultId: number;
  uniqueFaultId: string;
  locoNo?: string | null;
  faultCode?: string | null;
  whatsappNumber: string;
  message: string;
  status: AlertLogStatus;
  sentAt?: string | null;
  createdAt: string;
}

export interface ApiLog {
  id: number;
  endpoint: string;
  statusCode?: number | null;
  responseTime?: number | null;
  recordsFetched?: number | null;
  newRecords?: number | null;
  updatedRecords?: number | null;
  error?: string | null;
  createdAt: string;
}

export interface LoginHistory {
  id: number;
  userId: number;
  username: string;
  action: string;
  ip?: string | null;
  userAgent?: string | null;
  createdAt: string;
}

export interface Setting {
  key: string;
  value: string;
  updatedAt: string;
}

export interface SettingsMap {
  apiEndpoint: string;
  apiKey: string;
  refreshInterval: string;
  whatsappPhoneNumberId: string;
  whatsappAccessToken: string;
  whatsappRecipient: string;
  timezone: string;
  dashboardLink: string;
}

export interface UpdateSettingsRequest {
  apiEndpoint?: string;
  apiKey?: string;
  refreshInterval?: string;
  whatsappPhoneNumberId?: string;
  whatsappAccessToken?: string;
  whatsappRecipient?: string;
  timezone?: string;
  dashboardLink?: string;
}

export interface SchedulerStatus {
  running: boolean;
  lastRun?: string | null;
  nextRun?: string | null;
  intervalMinutes: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type Logout200 = {
  message: string;
};

export type GetDashboardStatsParams = {
/**
 * Date for stats (defaults to today)
 */
date?: string;
};

export type ListFaultsParams = {
page?: number;
pageSize?: number;
search?: string;
dateFrom?: string;
dateTo?: string;
zone?: string;
shed?: string;
locoNo?: string;
coachNumber?: string;
faultCode?: string;
category?: string;
moduleName?: string;
basicUnit?: string;
location?: string;
coachType?: string;
recoveryStatus?: string;
sortBy?: string;
sortOrder?: ListFaultsSortOrder;
};

export type ListFaultsSortOrder = typeof ListFaultsSortOrder[keyof typeof ListFaultsSortOrder];


export const ListFaultsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ExportFaultsParams = {
format: ExportFaultsFormat;
dateFrom?: string;
dateTo?: string;
zone?: string;
shed?: string;
category?: string;
recoveryStatus?: string;
};

export type ExportFaultsFormat = typeof ExportFaultsFormat[keyof typeof ExportFaultsFormat];


export const ExportFaultsFormat = {
  csv: 'csv',
  excel: 'excel',
  pdf: 'pdf',
} as const;

export type ExportFaults200 = {
  downloadUrl: string;
  filename: string;
  records: number;
};

export type GetFaultFilterOptions200 = {
  zones: string[];
  sheds: string[];
  categories: string[];
  faultCodes: string[];
  modules: string[];
  basicUnits: string[];
  locations: string[];
  coachTypes: string[];
};

export type GetCategoryPieChartParams = {
dateFrom?: string;
dateTo?: string;
};

export type GetFaultTrendParams = {
period?: GetFaultTrendPeriod;
dateFrom?: string;
dateTo?: string;
};

export type GetFaultTrendPeriod = typeof GetFaultTrendPeriod[keyof typeof GetFaultTrendPeriod];


export const GetFaultTrendPeriod = {
  hourly: 'hourly',
  daily: 'daily',
  monthly: 'monthly',
} as const;

export type GetFaultCodeChartParams = {
limit?: number;
dateFrom?: string;
dateTo?: string;
};

export type GetModuleChartParams = {
limit?: number;
dateFrom?: string;
dateTo?: string;
};

export type GetLocationChartParams = {
limit?: number;
dateFrom?: string;
dateTo?: string;
};

export type GetLocoChartParams = {
limit?: number;
dateFrom?: string;
dateTo?: string;
};

export type GetRecoveryTrendParams = {
period?: GetRecoveryTrendPeriod;
dateFrom?: string;
dateTo?: string;
};

export type GetRecoveryTrendPeriod = typeof GetRecoveryTrendPeriod[keyof typeof GetRecoveryTrendPeriod];


export const GetRecoveryTrendPeriod = {
  daily: 'daily',
  monthly: 'monthly',
} as const;

export type ListAlertsParams = {
page?: number;
pageSize?: number;
status?: ListAlertsStatus;
};

export type ListAlertsStatus = typeof ListAlertsStatus[keyof typeof ListAlertsStatus];


export const ListAlertsStatus = {
  sent: 'sent',
  failed: 'failed',
  pending: 'pending',
} as const;

export type ListAlerts200 = {
  data: AlertLog[];
  total: number;
};

export type ListApiLogsParams = {
page?: number;
pageSize?: number;
};

export type ListApiLogs200 = {
  data: ApiLog[];
  total: number;
};

export type ListLoginHistoryParams = {
page?: number;
pageSize?: number;
};

export type ListLoginHistory200 = {
  data: LoginHistory[];
  total: number;
};

export type TriggerFetch200 = {
  message: string;
  newRecords: number;
  updatedRecords: number;
};