# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indian Railways LocoNet Fault Monitoring System API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResult:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    UserInput:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]
      required:
        - username
        - email
        - password
        - role
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - username
        - email
        - role
        - createdAt
...
    ChartDataPoint:
      type: object
      properties:
        label:
          type: string
        value:
          type: integer
      required:
        - label
        - value
    TrendDataPoint:
      type: object
      properties:
        timestamp:
          type: string
        count:
          type: integer
      required:
        - timestamp
        - count
...
    ApiLog:
      type: object
      properties:
        id:
          type: integer
        endpoint:
          type: string
        statusCode:
          type: integer
          nullable: true
        responseTime:
          type: integer
          nullable: true
        recordsFetched:
          type: integer
          nullable: true
        newRecords:
          type: integer
          nullable: true
        updatedRecords:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - endpoint
        - createdAt
    LoginHistory:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        action:
          type: string
        ip:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - username
        - action
        - createdAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /auth/users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [auth]
      summary: Create user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /charts/category-pie:
    get:
      operationId: getCategoryPieChart
      tags: [charts]
      summary: Category-wise fault distribution
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Category chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
  /charts/fault-trend:
    get:
      operationId: getFaultTrend
      tags: [charts]
      summary: Fault trend over time (hourly/daily/monthly)
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [hourly, daily, monthly]
            default: daily
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendDataPoint"
  /charts/fault-code:
    get:
      operationId: getFaultCodeChart
      tags: [charts]
      summary: Top fault codes by count
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Fault code chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
  /charts/module:
    get:
      operationId: getModuleChart
      tags: [charts]
      summary: Module-wise fault distribution
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Module chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
  /charts/location:
    get:
      operationId: getLocationChart
      tags: [charts]
      summary: Location-wise fault distribution
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Location chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
  /charts/loco:
    get:
      operationId: getLocoChart
      tags: [charts]
      summary: Top locos by fault count
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Loco chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
  /charts/recovery-trend:
    get:
      operationId: getRecoveryTrend
      tags: [charts]
      summary: Recovery trend over time
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, monthly]
            default: daily
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Recovery trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendDataPoint"
...
  /audit/api-logs:
    get:
      operationId: listApiLogs
      tags: [audit]
      summary: List API fetch logs
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: API logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiLog"
                  total:
                    type: integer
                required:
                  - data
                  - total
  /audit/login-history:
    get:
      operationId: listLoginHistory
      tags: [audit]
      summary: List login history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Login history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/LoginHistory"
                  total:
                    type: integer
                required:
                  - data
                  - total
...